+++ src/server/data/mock-database.ts
@@ -1,29 +1,13 @@
-// Data Json Files
-const purchasesData = require('./purchases.json');
-const cheesesData = require('./cheeses.json');
+import { Cheese } from "./models/cheese.model";
+import { PurchaseHistory } from "./models/purchaseHistory.model";
+
+const purchasesData = require("./purchases.json");
+const cheesesData = require("./cheeses.json");
 const fs = require("fs");
 
-export interface Cheese {
-    id: number;
-    title: string;
-    price: number;
-    description: string;
-    category: string;
-    image: string;
-}
-
-export interface Order {
-    cheeseId: number;
-    quantity: number;
-}
-
-export interface PurchaseHistory {
-  id: string;
-  totalPrice: number;
-  orders: Order[]
-}
-
-// A mock database
+/**
+ * A mock-up database to get all cheeses and recent purchases.
+ */
 class MockDatabase {
     private cheeses: Cheese[];
     private purchaseHistory: PurchaseHistory[];
@@ -31,12 +15,12 @@ class MockDatabase {
     constructor() {
         this.refreshDatabase();
     }
-    
-    public get Cheeses() : Cheese[] {
-        return this.cheeses
+
+    public get Cheeses(): Cheese[] {
+        return this.cheeses;
     }
 
-    public get PurchaseHistory() : PurchaseHistory[] {
+    public get PurchaseHistory(): PurchaseHistory[] {
         return this.purchaseHistory;
     }
 
@@ -46,7 +30,10 @@ class MockDatabase {
     };
 
     public saveData = () => {
-        fs.writeFileSync("./src/server/data/purchases.json", JSON.stringify(this.purchaseHistory));
+        fs.writeFileSync(
+            "./src/server/data/purchases.json",
+            JSON.stringify(this.purchaseHistory)
+        );
     };
 }
