+++ src/server/data/models/cheese.model.ts
@@ -1,0 +1,8 @@
+export interface Cheese {
+    id: number;
+    title: string;
+    price: number;
+    description: string;
+    category: string;
+    image: string;
+}
