+++ src/server/data/models/purchaseHistory.model.ts
@@ -1,0 +1,10 @@
+export interface PurchaseHistory {
+    id: string;
+    totalPrice: number;
+    orders: Order[];
+}
+
+export interface Order {
+    cheeseId: number;
+    quantity: number;
+}
